requirements.txt
json == 2.0.9
Pillow == 8.4.0
torchvision == 0.10.0
torch == 1.10.0
